Problem 02 fix

+++ Problem_02.py
@@ -5,7 +5,7 @@ def calculate_fine(book_title, days_overdue, daily_rate=5.0, max_fine=150.0):
 
 book_title=input()
 days_overdue=int(input())
-daily_rate=int(input())
+daily_rate=float(input())
 fine=calculate_fine(book_title,days_overdue,daily_rate)
 print(f"Book: {book_title}")
 print(f"Days overdue: {days_overdue}")
